backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
python-multipart==0.0.6
Pillow==10.1.0
google-cloud-storage==2.10.0
googlemaps==4.10.0
